+++ query-engine/connectors/sql-query-connector/src/database/operations/write.rs
@@ -376,7 +376,7 @@ async fn generate_updates(
     record_filter: RecordFilter,
     args: WriteArgs,
     selected_fields: Option<&ModelProjection>,
-    limit: Option<i64>,
+    limit: Option<usize>,
     ctx: &Context<'_>,
 ) -> crate::Result<Vec<Query<'static>>> {
     if record_filter.has_selectors() {
@@ -399,7 +399,7 @@ pub(crate) async fn update_records(
     model: &Model,
     record_filter: RecordFilter,
     args: WriteArgs,
-    limit: Option<i64>,
+    limit: Option<usize>,
     ctx: &Context<'_>,
 ) -> crate::Result<usize> {
     if args.args.is_empty() {
@@ -421,7 +421,7 @@ pub(crate) async fn update_records_returning(
     record_filter: RecordFilter,
     args: WriteArgs,
     selected_fields: FieldSelection,
-    limit: Option<i64>,
+    limit: Option<usize>,
     ctx: &Context<'_>,
 ) -> crate::Result<ManyRecords> {
     let field_names: Vec<String> = selected_fields.db_names().collect();
@@ -458,7 +458,7 @@ pub(crate) async fn delete_records(
     conn: &dyn Queryable,
     model: &Model,
     record_filter: RecordFilter,
-    limit: Option<i64>,
+    limit: Option<usize>,
     ctx: &Context<'_>,
 ) -> crate::Result<usize> {
     let filter_condition = FilterBuilder::without_top_level_joins().visit_filter(record_filter.clone().filter, ctx);
@@ -474,7 +474,8 @@ pub(crate) async fn delete_records(
         {
             row_count += conn.execute(delete).await?;
             if let Some(old_remaining_limit) = remaining_limit {
-                let new_remaining_limit = old_remaining_limit - row_count as i64;
+                // u64 to usize cast here cannot 'overflow' as the number of rows was limited to MAX usize in the first place.
+                let new_remaining_limit = old_remaining_limit - row_count as usize;
                 if new_remaining_limit <= 0 {
                     break;
                 }
