+++ query-engine/connectors/sql-query-connector/src/database/transaction.rs
@@ -220,7 +220,7 @@ impl WriteOperations for SqlConnectorTransaction<'_> {
         model: &Model,
         record_filter: RecordFilter,
         args: WriteArgs,
-        limit: Option<i64>,
+        limit: Option<usize>,
         traceparent: Option<TraceParent>,
     ) -> connector::Result<usize> {
         let ctx = Context::new(&self.connection_info, traceparent);
@@ -237,7 +237,7 @@ impl WriteOperations for SqlConnectorTransaction<'_> {
         record_filter: RecordFilter,
         args: WriteArgs,
         selected_fields: FieldSelection,
-        limit: Option<i64>,
+        limit: Option<usize>,
         traceparent: Option<TraceParent>,
     ) -> connector::Result<ManyRecords> {
         let ctx = Context::new(&self.connection_info, traceparent);
@@ -283,7 +283,7 @@ impl WriteOperations for SqlConnectorTransaction<'_> {
         &mut self,
         model: &Model,
         record_filter: RecordFilter,
-        limit: Option<i64>,
+        limit: Option<usize>,
         traceparent: Option<TraceParent>,
     ) -> connector::Result<usize> {
         catch(&self.connection_info, async {
