+++ query-engine/core/src/query_ast/write.rs
@@ -368,7 +368,7 @@ pub struct UpdateManyRecords {
     /// Fields of updated records that client has requested to return.
     /// `None` if the connector does not support returning the updated rows.
     pub selected_fields: Option<UpdateManyRecordsFields>,
-    pub limit: Option<i64>,
+    pub limit: Option<usize>,
 }
 
 #[derive(Debug, Clone)]
@@ -398,7 +398,7 @@ pub struct DeleteRecordFields {
 pub struct DeleteManyRecords {
     pub model: Model,
     pub record_filter: RecordFilter,
-    pub limit: Option<i64>,
+    pub limit: Option<usize>,
 }
 
 #[derive(Debug, Clone)]
